move add_suffix() and process_path() to helpers.py
